requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pyTigerGraph>=1.0.0
pandas>=1.5.0
plotly>=5.15.0
numpy>=1.24.0
python-dotenv>=1.0.0